remove maze service
